[core] rename gl/*_buffer.hpp to gl/*_buffer_resource.hpp
