+++ src/core-files.json
@@ -517,7 +517,7 @@
         "mbgl/gl/extension.hpp": "src/mbgl/gl/extension.hpp",
         "mbgl/gl/features.hpp": "src/mbgl/gl/features.hpp",
         "mbgl/gl/framebuffer.hpp": "src/mbgl/gl/framebuffer.hpp",
-        "mbgl/gl/index_buffer.hpp": "src/mbgl/gl/index_buffer.hpp",
+        "mbgl/gl/index_buffer_resource.hpp": "src/mbgl/gl/index_buffer_resource.hpp",
         "mbgl/gl/object.hpp": "src/mbgl/gl/object.hpp",
         "mbgl/gl/program.hpp": "src/mbgl/gl/program.hpp",
         "mbgl/gl/program_binary_extension.hpp": "src/mbgl/gl/program_binary_extension.hpp",
@@ -529,7 +529,7 @@
         "mbgl/gl/value.hpp": "src/mbgl/gl/value.hpp",
         "mbgl/gl/vertex_array.hpp": "src/mbgl/gl/vertex_array.hpp",
         "mbgl/gl/vertex_array_extension.hpp": "src/mbgl/gl/vertex_array_extension.hpp",
-        "mbgl/gl/vertex_buffer.hpp": "src/mbgl/gl/vertex_buffer.hpp",
+        "mbgl/gl/vertex_buffer_resource.hpp": "src/mbgl/gl/vertex_buffer_resource.hpp",
         "mbgl/layout/clip_lines.hpp": "src/mbgl/layout/clip_lines.hpp",
         "mbgl/layout/layout.hpp": "src/mbgl/layout/layout.hpp",
         "mbgl/layout/merge_lines.hpp": "src/mbgl/layout/merge_lines.hpp",
